Fixed dependency-omission bug
GRNOC::RabbitMQ::Method depends on GRNOC::WebService::Regex, but the module
didn't explicitly include the module, nor was this dependency reflected in
the specfile for the package. The second omission led to the module bombing out
when I tried using it on a machine without GRNOC::WebService::Regex already
installed on it; the first caused bad things to happen even when
GRNOC::WebService::Regex was available on the system. This commit fixes both
problems.