Adding sync to prevent Dispatcher crash.
Theory:
Normally consume is called and its response is received before the bind request
goes on the wire. If however, a bind request hits the wire prior to consumeok
being received, the message queue inside AnyEvent::RabbitMQ may serve consumeok
to the queueok callback resulting in an error. I *think* messages on the wire
are assumed in-order in this case. Adding some synchronization to keep messages
in order seems to have fixed the issue.